requirements.txt
beautifulsoup4==4.9.3
blessed==1.17.6
boto3==1.18.36
botocore==1.21.36
bs4==0.0.1
cffi==1.14.6
cryptography==3.4.8
inquirer==2.7.0
jmespath==0.10.0
pycparser==2.20
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
wcwidth==0.2.5